+++ CHANGELOG.md
@@ -12,6 +12,7 @@ The following is a curated list of changes in the Enact limestone module, newest
 ### Changed
 
 - All components in `light` skin and high-contrast mode styling to match the latest GUI
+- `limestone/Card` to display `captions` on focus
 - `limestone/CheckboxItem` styling to match the latest GUI
 - `limestone/FlexiblePopupPanels` styling to match the latest GUI
 - `limestone/FormCheckboxItem` styling to match the latest GUI
@@ -23,6 +24,7 @@ The following is a curated list of changes in the Enact limestone module, newest
 ### Fixed
 
 - `limestone/Card` for `focusRing` mode
+- `limestone/Card` `background` for captions
 - `limestone/Popup` to restore focus properly when popup hides
 - `limestone/VideoPlayer` size of the icons from the control buttons